2.17: changed ring of force's armed boost to use 64-bit integers

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfForce.java
@@ -51,7 +51,7 @@ public class RingOfForce extends Ring {
 		return new Force();
 	}
 	
-	public static int armedDamageBonus( Char ch ){
+	public static long armedDamageBonus( Char ch ){
 		return getBuffedBonus( ch, Force.class);
 	}
